+++ frontend/app/src/entities/permission/constants.ts
@@ -1,0 +1,8 @@
+import { Permission } from "@/entities/permission/types";
+
+export const PERMISSION_ALLOW_ALL: Permission = {
+  create: { isAllowed: true },
+  view: { isAllowed: true },
+  update: { isAllowed: true },
+  delete: { isAllowed: true },
+};
